TickTock-NPULibrary/requirements.txt
kornia==0.6.3
matplotlib==3.3.3
numpy==1.19.4
opencv_python==4.12.0.88
pandas==1.3.5
Pillow==9.5.0
Pillow==11.3.0
Requests==2.32.5
scipy==1.7.1
torch==1.10.0+cu113
tqdm==4.59.0
